+++ src/GSvar/CnvWidget.h
@@ -57,8 +57,9 @@ private slots:
 	///Flags all unfiltered (=currently visible) CNVs as artefacts in somatic report configuration
 	void flagVisibleSomaticCnvsAsArtefacts();
 
-	///Set custom column widths
-	void adaptColumnWidthsCustom();
+	void openColumnSettings();
+	void adaptColumnWidthsAndHeights();
+	void showAllColumns();
 
 private:
 	void initGUI();