+++ src/GSvar/MainWindow.cpp
@@ -155,6 +155,8 @@ QT_CHARTS_USE_NAMESPACE
 #include "CustomProxyService.h"
 #include "GeneInterpretabilityDialog.h"
 #include "HerediVarImportDialog.h"
+#include "Background/IGVInitCacheWorker.h"
+
 
 MainWindow::MainWindow(QWidget *parent)
 	: QMainWindow(parent)
@@ -167,7 +169,7 @@ MainWindow::MainWindow(QWidget *parent)
 	, variants_changed_()
 	, last_report_path_(QDir::homePath())
 	, init_timer_(this, true)
-	, server_version_()
+	, server_version_()   
 {
     // Automatic configuration will be triggered, if a template file is detected and no settings files are present.
     // A new settings.ini file is created with parameters based on the current application path value. If there is no
@@ -294,7 +296,6 @@ MainWindow::MainWindow(QWidget *parent)
 
 	//variants tool bar
 	connect(ui_.vars_copy_btn, SIGNAL(clicked(bool)), ui_.vars, SLOT(copyToClipboard()));
-	connect(ui_.vars_resize_btn, SIGNAL(clicked(bool)), ui_.vars, SLOT(adaptColumnWidthsCustom()));
 	ui_.vars_export_btn->setMenu(new QMenu());
 	ui_.vars_export_btn->menu()->addAction("Export GSvar (filtered)", this, SLOT(exportGSvar()));
 	ui_.vars_export_btn->menu()->addAction("Export VCF (filtered)", this, SLOT(exportVCF()));
@@ -320,6 +321,10 @@ MainWindow::MainWindow(QWidget *parent)
 	ui_.vars_af_hist->menu()->addSeparator();
 	ui_.vars_af_hist->menu()->addAction("Show CN histogram (in given region)", this, SLOT(showCnHistogram()));
 	ui_.vars_af_hist->menu()->addAction("Show BAF histogram (in given region)", this, SLOT(showBafHistogram()));
+	ui_.vars_resize_btn->setMenu(new QMenu());
+	ui_.vars_resize_btn->menu()->addAction("Open column settings", this, SLOT(openColumnSettings()));
+	ui_.vars_resize_btn->menu()->addAction("Apply column width settings", ui_.vars, SLOT(adaptColumnWidths()));
+	ui_.vars_resize_btn->menu()->addAction("Show all columns", ui_.vars, SLOT(showAllColumns()));
 
 	connect(ui_.ps_details, SIGNAL(clicked(bool)), this, SLOT(openProcessedSampleTabsCurrentAnalysis()));
 
@@ -402,7 +407,7 @@ MainWindow::MainWindow(QWidget *parent)
 		{
 			Log::error("Could not set CURL_CA_BUNDLE variable, access to BAM files over HTTPS may not be possible");
 		}
-	}
+	}   
 }
 
 QString MainWindow::appName() const
@@ -445,6 +450,13 @@ bool MainWindow::isServerRunning()
 	}
 
     return true;
+}
+
+void MainWindow::lazyLoadIGVfiles()
+{
+    IgvSessionManager::get(0).removeCache();
+    Log::info("Started loading file location information needed for the IGV initialization");
+    IgvSessionManager::get(0).startCachingForRegularIGV(variants_.type(), filename_);
 }
 
 void MainWindow::checkServerAvailability()
@@ -744,7 +756,7 @@ void MainWindow::on_actionImportVariants_triggered()
 
 void MainWindow::on_actionIgvClear_triggered()
 {
-	IgvSessionManager::get(0).clear();
+    IgvSessionManager::get(0).clear();
 }
 
 void MainWindow::on_actionIgvDocumentation_triggered()
@@ -792,7 +804,7 @@ void MainWindow::on_actionSV_triggered()
 		int passing_vars = filter_result_.countPassing();
 		if (passing_vars>3000)
 		{
-			int res = QMessageBox::question(this, "Continue?", "There are " + QString::number(passing_vars) + " variants that pass the filters.\nGenerating the list of candidate genes for compound-heterozygous hits may take very long for this amount of variants.\nDo you want to continue?", QMessageBox::Yes, QMessageBox::No);
+			int res = QMessageBox::question(this, "Continue?", "There are " + QString::number(passing_vars) + " small variants that pass the filters.\nGenerating the list of candidate genes for compound-heterozygous hits may take very long for this amount of variants.\nDo you want to continue?", QMessageBox::Yes, QMessageBox::No);
 			if(res==QMessageBox::No) return;
 		}
 		for (int i=0; i<variants_.count(); ++i)
@@ -822,13 +834,10 @@ void MainWindow::on_actionSV_triggered()
 		SvWidget* sv_widget;
 		if(svs_.isSomatic())
 		{
-			QString ps_id =	"";
-
-			// somatic
-			sv_widget = new SvWidget(this, svs_, ps_id, somatic_report_settings_.report_config, het_hit_genes);
+			sv_widget = new SvWidget(this, svs_, somatic_report_settings_.report_config, het_hit_genes);
 			connect(sv_widget, SIGNAL(updateSomaticReportConfiguration()), this, SLOT(storeSomaticReportConfig()));
 		}
-		else
+		else //germline
 		{
 			//determine processed sample ID (needed for report config)
 			QString ps_id = "";
@@ -840,13 +849,11 @@ void MainWindow::on_actionSV_triggered()
 			}
 
 			//open SV widget
-			sv_widget = new SvWidget(this, svs_, ps_id, svs_.isSomatic() ? nullptr : report_config, het_hit_genes);
+			sv_widget = new SvWidget(this, svs_, ps_id, report_config, het_hit_genes);
 		}
 
 		auto dlg = GUIHelper::createDialog(sv_widget, "Structural variants of " + variants_.analysisName());
 		addModelessDialog(dlg);
-
-
 	}
 	catch(FileParseException error)
 	{
@@ -882,7 +889,7 @@ void MainWindow::on_actionCNV_triggered()
 		int passing_vars = filter_result_.countPassing();
 		if (passing_vars>3000)
 		{
-			int res = QMessageBox::question(this, "Continue?", "There are " + QString::number(passing_vars) + " variants that pass the filters.\nGenerating the list of candidate genes for compound-heterozygous hits may take very long for this amount of variants.\nPlease set a filter for the variant list, e.g. the recessive filter, and retry!\nDo you want to continue?", QMessageBox::Yes, QMessageBox::No);
+			int res = QMessageBox::question(this, "Continue?", "There are " + QString::number(passing_vars) + " small variants that pass the filters.\nGenerating the list of candidate genes for compound-heterozygous hits may take very long for this amount of variants.\nPlease set a filter for the variant list, e.g. the recessive filter, and retry!\nDo you want to continue?", QMessageBox::Yes, QMessageBox::No);
 			if(res==QMessageBox::No) return;
 		}
 		for (int i=0; i<variants_.count(); ++i)
@@ -1691,7 +1698,7 @@ void MainWindow::delayedInitialization()
 				{
 					openSampleFromNGSD(arg);
 				}
-			}
+            }
 		}
 		else
 		{
@@ -2060,7 +2067,14 @@ void MainWindow::on_actionEncrypt_triggered()
 
 void MainWindow::on_actionSettings_triggered()
 {
+	openSettingsDialog();
+}
+
+void MainWindow::openSettingsDialog(QString page_name, QString section)
+{
 	SettingsDialog dlg(this);
+	dlg.setWindowFlags(Qt::Window);
+	dlg.gotoPage(page_name, section);
 	if (dlg.exec()==QDialog::Accepted)
 	{
 		dlg.storeSettings();
@@ -2085,6 +2099,11 @@ void MainWindow::on_actionRunOverview_triggered()
 
 	SequencingRunOverview* widget = new SequencingRunOverview(this);
 	openTab(QIcon(":/Icons/NGSD_run_overview.png"), name, type, widget);
+}
+
+void MainWindow::openColumnSettings()
+{
+	openSettingsDialog("columns", variantTypeToString(VariantType::SNVS_INDELS));
 }
 
 void MainWindow::addModelessDialog(QSharedPointer<QDialog> dlg, bool maximize)
@@ -2241,7 +2260,7 @@ void MainWindow::openProcessedSampleFromNGSD(QString processed_sample_name, bool
 			file = analysis_info_list[index].analysis_file;
 		}
 
-		loadFile(file);
+        loadFile(file);
 	}
 	catch (Exception& e)
 	{
@@ -2271,7 +2290,7 @@ void MainWindow::openSampleFromNGSD(QString sample_name)
 			if (!ok) return;
 
 			openProcessedSampleFromNGSD(ps, false);
-		}
+        }
 	}
 	catch (Exception& e)
 	{
@@ -2515,20 +2534,32 @@ int MainWindow::openTab(QIcon icon, QString name, TabType type, QWidget* widget)
 
 void MainWindow::closeTab(int index)
 {
+	//main variant list
 	if (index==0)
 	{
-		int res = QMessageBox::question(this, "Close file?", "Do you want to close the current sample?", QMessageBox::Yes, QMessageBox::No);
-		if (res==QMessageBox::Yes)
+		if (filename_!="")
 		{
-			loadFile();
+			int res = QMessageBox::question(this, "Close file?", "Do you want to close the current sample?", QMessageBox::Yes, QMessageBox::No);
+			if (res==QMessageBox::Yes)
+			{
+				loadFile();
+			}
 		}
+		return;
 	}
-	else
+
+	//for analysis status widget and refresh is done > abort
+	QWidget* widget = ui_.tabs->widget(index);
+	AnalysisStatusWidget* analysis_status_widget = widget->findChild<AnalysisStatusWidget*>();
+	while (analysis_status_widget!=nullptr && analysis_status_widget->updateIsRunning())
 	{
-		QWidget* widget = ui_.tabs->widget(index);
-		ui_.tabs->removeTab(index);
-		widget->deleteLater();
+		QMessageBox::information(this, "Analysis status tab", "Please wait until table update is done before closing the analysis status tab!");
+		return;
 	}
+
+	//remove tab and delete tab widget
+	ui_.tabs->removeTab(index);
+	widget->deleteLater();
 }
 
 bool MainWindow::focusTab(TabType type, QString name)
@@ -2576,6 +2607,7 @@ void MainWindow::loadFile(QString filename, bool show_only_error_issues)
 	variants_changed_.clear();
 	cnvs_.clear();
 	svs_.clear();
+	res_.clear();
 	ui_.vars->clearContents();
 	report_settings_ = ReportSettings();
 	connect(report_settings_.report_config.data(), SIGNAL(variantsChanged()), this, SLOT(storeReportConfig()));
@@ -2605,6 +2637,7 @@ void MainWindow::loadFile(QString filename, bool show_only_error_issues)
 			GlobalServiceProvider::setFileLocationProvider(QSharedPointer<FileLocationProviderLocal>(new FileLocationProviderLocal(filename, variants_.getSampleHeader(), variants_.type())));
 			mode_title = " (local mode)";
 		}
+        lazyLoadIGVfiles();
 
 		//load CNVs
 		timer.restart();
@@ -5788,9 +5821,19 @@ void MainWindow::showBackgroundJobDialog()
 	bg_job_dialog_->show();
 }
 
-void MainWindow::startJob(BackgroundWorkerBase* worker, bool show_busy_dialog)
+int MainWindow::startJob(BackgroundWorkerBase* worker, bool show_busy_dialog)
 {
-	bg_job_dialog_->start(worker, show_busy_dialog);
+    return bg_job_dialog_->start(worker, show_busy_dialog);
+}
+
+QString MainWindow::getJobStatus(int id)
+{
+    return bg_job_dialog_->getJobStatus(id);
+}
+
+QString MainWindow::getJobMessages(int id)
+{
+    return bg_job_dialog_->getJobMessages(id);
 }
 
 void MainWindow::on_actionVirusDetection_triggered()
@@ -5889,7 +5932,7 @@ void MainWindow::editVariantClassification(VariantList& variants, int index, boo
 			if(!clinvar_class.isEmpty() && clinvar_class!=new_class)
 			{
 				//update on ClinVar
-				int return_value = QMessageBox::information(this, "Clinvar upload required!", "Variant already uploaded to ClinVar. You should also update the classification there!", QMessageBox::Ok, QMessageBox::NoButton);
+                int return_value = QMessageBox::information(this, "Clinvar upload required!", "Variant already uploaded to ClinVar. You should also update the classification there!", QMessageBox::Ok, QMessageBox::NoButton);
 				if(return_value == QMessageBox::Ok)	uploadToClinvar(index);
 			}
 		}