+++ src/GSvar/MainWindow.h
@@ -2,7 +2,6 @@
 #define MAINWINDOW_H
 
 #include <QMainWindow>
-#include "Background/BackgroundJobDialog.h"
 #include "ui_MainWindow.h"
 #include "VariantList.h"
 #include "BedFile.h"
@@ -20,6 +19,7 @@
 #include "ClickableLabel.h"
 #include "ImportDialog.h"
 #include "RepeatLocusList.h"
+#include "Background/BackgroundJobDialog.h"
 
 ///Tab type
 enum class TabType
@@ -49,6 +49,9 @@ public:
 
 	/// Gets server API information to make sure ther the server is currently running
 	bool isServerRunning();
+
+    /// Gets the list of files needed to initialize IGV, reduces the delay before the fist call to IGV
+    void lazyLoadIGVfiles();
 
 	///Returns the result of applying filters to the variant list
 	void applyFilters(bool debug_time);
@@ -386,6 +389,10 @@ public slots:
 	void on_actionSampleSearch_triggered();
 	///Show run overview
 	void on_actionRunOverview_triggered();
+	///Open column settings dialog
+	void openColumnSettings();
+	///Open settings dialog on a specific page
+	void openSettingsDialog(QString page_name="general", QString section = "");
 
 	///Subpanel design dialog
 	void openSubpanelDesignDialog(const GeneSet& genes = GeneSet());
@@ -468,8 +475,12 @@ public slots:
     void changeIgvIconToNormal();
 	//Open background jobs dialog
 	void showBackgroundJobDialog();
-	//Starts a background job
-	void startJob(BackgroundWorkerBase* worker, bool show_busy_dialog);
+    //Starts a background job and returns its id
+    int startJob(BackgroundWorkerBase* worker, bool show_busy_dialog);
+    //Returns information about a background job status by its id
+    QString getJobStatus(int id);
+    //Returns error messages for a background job by its id (if it failed)
+    QString getJobMessages(int id);
 
     ///close the app and logout (if in client-sever mode)
 	void closeAndLogout();
@@ -545,8 +556,7 @@ private:
 	QString displayed_maintenance_message_id_;
 
     //current server version (if in client-server mode)
-    QString server_version_;
-
+    QString server_version_; 
 };
 
 #endif // MAINWINDOW_H